Renamed include dir from API to RayTraceEngine
