+++ include/RayTraceEngine/BasicStructures.h
