+++ include/RayTraceEngine/Object.h
