+++ src/CMakeLists.txt
@@ -5,7 +5,7 @@ set_target_properties(RayTraceEngine PROPERTIES VERSION ${PROJECT_VERSION} SOVER
 
 # For MacOS Framework
 if(${CMAKE_SYSTEM_NAME} MATCHES "Darwin")
-    set_target_properties(RayTraceEngine PROPERTIES PUBLIC_HEADER "include/API/RayTraceCore.h;include/API/RayEngine.h;include/API/MissShader.h;include/API/HitShader.h;include/API/OcclusionShader.h;include/API/PierceShader.h;include/API/RayGeneratorShader.h;include/API/Object.h;include/API/Pipeline.h;include/API/Shader.h;include/API/BasicStructures.h;include/API/TriangleMeshObject.h")
+    set_target_properties(RayTraceEngine PROPERTIES PUBLIC_HEADER "include/RayTraceEngine/RayTraceCore.h;include/RayTraceEngine/RayEngine.h;include/RayTraceEngine/MissShader.h;include/RayTraceEngine/HitShader.h;include/RayTraceEngine/OcclusionShader.h;include/RayTraceEngine/PierceShader.h;include/RayTraceEngine/RayGeneratorShader.h;include/RayTraceEngine/Object.h;include/RayTraceEngine/Pipeline.h;include/RayTraceEngine/Shader.h;include/RayTraceEngine/BasicStructures.h;include/RayTraceEngine/TriangleMeshObject.h")
 endif()
 
 configure_file(RayTraceEngine.pc.in RayTraceEngine.pc @ONLY)
